+++ src/main/java/br/com/techpib/ap/ms_anotacoes_ebd/core/data/form/AnotacaoForm.java
@@ -12,7 +12,7 @@ public class AnotacaoForm {
     @NotNull
     private UUID idUsuario;
 
-    private Integer sequencialAnotacao;
+    private Long sequencialAnotacao;
 
     @Size(max = 100)
     private String titulo;
